requirements-base.txt
# 重いライブラリ
pyyaml
numpy
pandas
matplotlib
seaborn
scikit-learn
torch
torchvision
